minor fixes to preprocess

+++ preprocess_rgb.py
@@ -48,34 +48,35 @@ def make_subclips():
 def make_npy():
     os.chdir(DATASET)
     for s in os.listdir(DATASET):
-        print(s)
-        cap = cv.VideoCapture(s)
-        vid = VideoFileClip(s)
-        num_frames = math.ceil(FRAME_RATE * vid.end)
-        frame_no = 0
-        final_input = np.zeros((1, num_frames, IMG_SIZE, IMG_SIZE, 3))
+        if not os.path.isfile(s[-4] + ".npy"):
+            print(s)
+            cap = cv.VideoCapture(s)
+            vid = VideoFileClip(s)
+            num_frames = math.ceil(FRAME_RATE * vid.end)
+            frame_no = 0
+            final_input = np.zeros((1, num_frames, IMG_SIZE, IMG_SIZE, 3))
 
-        while frame_no < num_frames:
-            ret, frame = cap.read()
-            cv.imshow('fr', frame)
-            old_height, old_width = frame.shape[0], frame.shape[1]
-            dim = (old_width * 256//old_height, 256)
-            frame = cv.resize(frame, dim, interpolation=cv.INTER_LINEAR)
-            center = (frame.shape[0] // 2, frame.shape[1] // 2)
-            x = center[1] - IMG_SIZE // 2
-            y = center[0] - IMG_SIZE // 2
-            frame = frame[y:y+IMG_SIZE, x:x+IMG_SIZE]
-            cv.imshow("fr", frame)
-            norm_pixels = np.zeros((3,))
-            for i in range(IMG_SIZE):
-                for j in range(IMG_SIZE):
-                    pixel = frame[i, j][:]
-                    for k in range(len(pixel)):
-                        norm = pixel[k]/255 * 2 - 1
-                        norm_pixels[k] = norm
-                    final_input[0][frame_no][i][j] = norm_pixels
-            frame_no += 1
-        np.save(s[:-4], final_input)
+            while frame_no < num_frames:
+                ret, frame = cap.read()
+                cv.imshow('fr', frame)
+                old_height, old_width = frame.shape[0], frame.shape[1]
+                dim = (old_width * 256//old_height, 256)
+                frame = cv.resize(frame, dim, interpolation=cv.INTER_LINEAR)
+                center = (frame.shape[0] // 2, frame.shape[1] // 2)
+                x = center[1] - IMG_SIZE // 2
+                y = center[0] - IMG_SIZE // 2
+                frame = frame[y:y+IMG_SIZE, x:x+IMG_SIZE]
+                cv.imshow("fr", frame)
+                norm_pixels = np.zeros((3,))
+                for i in range(IMG_SIZE):
+                    for j in range(IMG_SIZE):
+                        pixel = frame[i, j][:]
+                        for k in range(len(pixel)):
+                            norm = pixel[k]/255 * 2 - 1
+                            norm_pixels[k] = norm
+                        final_input[0][frame_no][i][j] = norm_pixels
+                frame_no += 1
+            np.save(s[:-4], final_input)
 
 
 if __name__ == "__main__":
